Move "stats" to networkProvider.
